+++ src/core/support/proc/mod.rs
@@ -1,5 +1,6 @@
 mod arch;
 mod base;
 mod output;
+mod stream;
 
-pub use arch::{Output, Proc};
+pub use arch::{Output, Proc, Recv, Stream};